+++ test/src/unit/unit_cdbn_2.cpp
@@ -33,7 +33,7 @@ TEST_CASE("unit/cdbn/mnist/7", "[cdbn][ap][svm][unit]") {
             dll::conv_rbm_desc<10, 10, 7, 20, 3, 3, dll::momentum, dll::batch_size<25>>::layer_t,
             dll::avgp_layer_3d_desc<20, 8, 5, 2, 1, 1>::layer_t>>::dbn_t dbn_t;
 
-    auto dataset = mnist::read_dataset_direct<std::vector, etl::fast_dyn_matrix<double, 1, 28, 28>>(200);
+    auto dataset = mnist::read_dataset_direct<std::vector, etl::fast_dyn_matrix<float, 1, 28, 28>>(200);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
@@ -62,7 +62,7 @@ TEST_CASE("unit/cdbn/mnist/8", "[cdbn][ap][svm][unit]") {
             dll::conv_rbm_desc<1, 28, 28, 20, 15, 17, dll::momentum, dll::batch_size<25>>::layer_t,
             dll::conv_rbm_desc<20, 14, 12, 20, 7, 3, dll::momentum, dll::batch_size<25>>::layer_t, dll::mp_layer_3d_desc<20, 8, 10, 1, 1, 1>::layer_t, dll::avgp_layer_3d_desc<20, 8, 10, 1, 1, 1>::layer_t>>::dbn_t dbn_t;
 
-    auto dataset = mnist::read_dataset_direct<std::vector, etl::fast_dyn_matrix<double, 1, 28, 28>>(100);
+    auto dataset = mnist::read_dataset_direct<std::vector, etl::fast_dyn_matrix<float, 1, 28, 28>>(100);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
@@ -91,7 +91,7 @@ TEST_CASE("unit/cdbn/mnist/9", "[dbn][conv][mnist][patches][unit]") {
             dll::conv_rbm_desc_square<1, 14, 10, 5, dll::parallel_mode, dll::momentum, dll::batch_size<10>>::layer_t,
             dll::conv_rbm_desc_square<10, 10, 10, 5, dll::parallel_mode, dll::momentum, dll::batch_size<10>>::layer_t>>::dbn_t dbn_t;
 
-    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<double, 3>>(50);
+    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<float, 3>>(50);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
@@ -112,7 +112,7 @@ TEST_CASE("unit/cdbn/mnist/10", "[dbn][conv][mnist][patches][memory][unit]") {
             dll::conv_rbm_desc_square<20, 10, 20, 5, dll::momentum, dll::batch_size<10>>::layer_t>,
         dll::batch_mode>::dbn_t dbn_t;
 
-    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<double, 3>>(50);
+    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<float, 3>>(50);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
@@ -139,7 +139,7 @@ TEST_CASE("unit/cdbn/mnist/11", "[dbn][conv][mnist][patches][unit]") {
             dll::conv_rbm_desc_square<1, 14, 20, 5, dll::momentum, dll::batch_size<10>>::layer_t,
             dll::conv_rbm_desc_square<20, 10, 20, 5, dll::momentum, dll::batch_size<10>>::layer_t>>::dbn_t dbn_t;
 
-    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<double, 3>>(50);
+    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<float, 3>>(50);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
@@ -159,7 +159,7 @@ TEST_CASE("unit/cdbn/mnist/12", "[dbn][conv][mnist][patches][unit]") {
             dll::dyn_conv_rbm_desc<dll::momentum>::layer_t,
             dll::dyn_conv_rbm_desc<dll::momentum>::layer_t>>::dbn_t dbn_t;
 
-    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<double, 3>>(50);
+    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<float, 3>>(50);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
